requirements.txt
python-telegram-bot>=20.0
requests>=2.25.0
gspread>=5.0.0
python-dotenv>=0.21.0
selenium>=4.0.0
twocaptcha>=0.0.1
playwright>=1.30.0
